Refactoring to reduce duplication. Removing trailing whitespace.
